Update loading state and UI adjustments on Todo Page
This revision introduces a loading state for the Todo Page and adjusts the UI. Now, when the todos are loading, a CircularProgressIndicator is shown in the middle of the screen instead of a blank page. In addition, the margin for each todo item inside the ListView.builder is reduced for better screen usage. Also, the unnecessary debug banner on the home screen is removed.